+++ packages/client/mutations/toasts/popNotificationToast.ts
@@ -7,7 +7,6 @@ import {Snack} from '../../components/Snackbar'
 import {OnNextHandler, OnNextHistoryContext} from '../../types/relayMutations'
 import SetNotificationStatusMutation from '../SetNotificationStatusMutation'
 import mapDiscussionMentionedToToast from './mapDiscussionMentionedToToast'
-import mapKudosReceivedToToast from './mapKudosReceivedToToast'
 import mapMentionedToToast from './mapMentionedToToast'
 import mapPromptToJoinOrgToToast from './mapPromptToJoinOrgToToast'
 import mapRequestToJoinOrgToToast from './mapRequestToJoinOrgToToast'
@@ -26,8 +25,7 @@ const typePicker: Partial<
   TEAMS_LIMIT_EXCEEDED: mapTeamsLimitExceededToToast,
   TEAMS_LIMIT_REMINDER: mapTeamsLimitReminderToToast,
   PROMPT_TO_JOIN_ORG: mapPromptToJoinOrgToToast,
-  REQUEST_TO_JOIN_ORG: mapRequestToJoinOrgToToast,
-  KUDOS_RECEIVED: mapKudosReceivedToToast
+  REQUEST_TO_JOIN_ORG: mapRequestToJoinOrgToToast
 }
 
 graphql`
@@ -43,7 +41,6 @@ graphql`
       ...mapTeamsLimitReminderToToast_notification @relay(mask: false)
       ...mapPromptToJoinOrgToToast_notification @relay(mask: false)
       ...mapRequestToJoinOrgToToast_notification @relay(mask: false)
-      ...mapKudosReceivedToToast_notification @relay(mask: false)
     }
   }
 `