+++ packages/server/dataloader/primaryKeyLoaderMakers.ts
@@ -1,7 +1,6 @@
 import getKysely from '../postgres/getKysely'
 import {getDiscussionsByIds} from '../postgres/queries/getDiscussionsByIds'
 import {getDomainJoinRequestsByIds} from '../postgres/queries/getDomainJoinRequestsByIds'
-import {getKudosesByIds} from '../postgres/queries/getKudosesByIds'
 import getMeetingSeriesByIds from '../postgres/queries/getMeetingSeriesByIds'
 import getMeetingTemplatesByIds from '../postgres/queries/getMeetingTemplatesByIds'
 import {getTeamPromptResponsesByIds} from '../postgres/queries/getTeamPromptResponsesByIds'
@@ -20,7 +19,6 @@ export const teamPromptResponses = primaryKeyLoaderMaker(getTeamPromptResponsesB
 export const meetingSeries = primaryKeyLoaderMaker(getMeetingSeriesByIds)
 export const meetingTemplates = primaryKeyLoaderMaker(getMeetingTemplatesByIds)
 export const domainJoinRequests = primaryKeyLoaderMaker(getDomainJoinRequestsByIds)
-export const kudoses = primaryKeyLoaderMaker(getKudosesByIds)
 
 export const embeddingsMetadata = primaryKeyLoaderMaker((ids: readonly number[]) => {
   return getKysely().selectFrom('EmbeddingsMetadata').selectAll().where('id', 'in', ids).execute()